update dpo example
