+++ llm_tricks/DPO_example/dataset.py
@@ -1,4 +1,3 @@
-import torch
 from torch.utils.data import Dataset
 import json
