+++ llm_tricks/DPO_example/evaluate.py
@@ -1,0 +1,32 @@
+import torch
+from loss import compute_loss_dataloader
+
+
+def evaluate_loss_dataloader(policy_model, reference_model, train_loader, val_loader, beta, eval_iter):
+    policy_model.eval()
+    with torch.no_grad():
+        train_loss, train_chosen_rewards, train_rejected_rewards = compute_loss_dataloader(
+            data_loader=train_loader,
+            policy_model=policy_model,
+            reference_model=reference_model,
+            beta=beta,
+            num_batches=eval_iter
+        )
+        val_loss, val_chosen_rewards, val_rejected_rewards = compute_loss_dataloader(
+            data_loader=val_loader,
+            policy_model=policy_model,
+            reference_model=reference_model,
+            beta=beta,
+            num_batches=eval_iter
+        )
+    res = {
+        "train_loss": train_loss,
+        "train_chosen_reward": train_chosen_rewards,
+        "train_rejected_reward": train_rejected_rewards,
+        "val_loss": val_loss,
+        "val_chosen_reward": val_chosen_rewards,
+        "val_rejected_reward": val_rejected_rewards
+    }
+
+    policy_model.train()
+    return res
